Fix behaviour of textEntry and entry controls on OSX.
See https://github.com/jodonoghue/wxHaskell/issues/1#issuecomment-5202890

It seems that using the wxTE_RICH attribute in textEntry and entry causes a multi-line text control style to be created by default on OS X. This is new behaviour on wxWidgets 2.9.x.

This patch makes the wxTE_RICH and wxTE_RICH2 attribute settings dependent on the host OS (only sets them on Windows)

This patch also fixes bugs associated with the use of textCtrlClear followed by textCtrlWriteText, reported in the same issue.

Note that this does not fix the problem that the textColor attribute is not respected.