Use the same custom hook for `cabal copy`
It seems like `cabal` uses `copy` rather than `install` when installing
inside of a sandbox from a package linked in using `add-source`. This
means that `myInstHook` is never called when running
`cabal install wxc` inside of a sandbox and so the shared library is
not properly copied over to the sandbox's wxc lib folder.